+++ src/main/kotlin/process/pitch/OpenUtauPitchConversion.kt
@@ -4,7 +4,6 @@ import kotlin.math.roundToInt
 import model.Note
 import model.Pitch
 import model.Tempo
-import process.TickTimeTransformer
 import process.interpolateCosineEaseIn
 import process.interpolateCosineEaseInOut
 import process.interpolateCosineEaseOut
