+++ src/test/kotlin/process/pitch/TickTimeTransformerTest.kt
@@ -1,4 +1,4 @@
-package process
+package process.pitch
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
